earthing-report-generator/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Calculations - NumPy 1.x for stability
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# RAG and embeddings
huggingface-hub==0.16.4
transformers==4.30.2
sentence-transformers==2.2.2
chromadb==0.4.18

# LLM client
anthropic==0.7.8
python-dotenv==1.0.0

# ChromaDB dependencies (explicit versions to avoid conflicts)
chroma-hnswlib==0.7.3
onnxruntime==1.16.3
tokenizers==0.15.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.5